auth-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==1.4.49
pydantic==1.10.7
python-dotenv==1.0.0
alembic==1.10.4
python-dateutil==2.8.2
pyjwt==2.8.0
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.0.0.post2
